Adding files of the completed project
This is what the group project finished. This is a client server architecture application. Originally this application was going to have a web based front end using REACT but then the project got objectives allowed for it to just be a console application. See the README.md file to get more details about the project and who was part of the group.